news_system/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
mysql-connector-python==8.1.0
Pillow==10.0.0
email-validator==2.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1 